fetat EVE-13 bug fixed and added validationToCreateEvent in EventValidator class

+++ src/main/java/pl/envelo/melo/domain/event/dto/NewEventDto.java
@@ -42,7 +42,7 @@ public class NewEventDto {
     private PeriodicType periodicType;
     @Size(max = 100)
     private Set<HashtagDto> hashtags;
-    private int memberLimit;
+    private long memberLimit;
     private Set<Integer> invitedMembers;
     private Integer unitId;
     private LocationDto location;
